requirements.txt
PyPDF2
pdf2image
pytesseract
pillow
python-dotenv